==> VI_data_collection/4_machine_dimension_analysis/6_machine_factor_comparison_1.0.Rmd <==
---
title: "machine dimension exploration"
---

```{r}
library('here')            # file logistics
library('data.table')      # data manipulation
library('tidyverse')       # data manipulation
library("psych")           # factor analysis
library("ggplot2")         # visualization
library("corrplot")        # correlation plot
library("ggridges")        # joyplots
library('patchwork')       # multiple plots 

options(scipen=999)
theme_set(theme_minimal())

values <- c("POWER", "ACHIEVEMENT", "HEDONISM", 
            "STIMULATION", "SELF", "UNIVERSALISM", 
            "BENEVOLENCE", "TRADITION", 
            "CONFORMITY", "SECURITY")

load(file = here("VI_data_collection", "_data","intermediary_data", "machines_as_fixed_mean_dfs.RDS"))
load(file = here("VI_data_collection", "_data","intermediary_data", "machines_as_fixed_factor_dfs.RDS"))

rm(machines_as_subjects_df)
```

```{r}
# subset datasets
machines_factors <- machines_as_fixed_factor_dfs$machines_as_subjects_a
machines_means   <- machines_as_fixed_mean_dfs$machines_as_subjects_a
```

Here I examine the factors comprised of all the machine setups and the means. 

```{r}
cors <-lapply(values, function(x){
  #x<-values[2]
  
  # made df of columns that were averaged
  means_df <- machines_means %>% select(starts_with(x)) %>%
    select(contains("mean"))
  
  # make df of the factors
  factors_df <- machines_factors %>% select(starts_with(x)) %>%
    select(contains("fa"))
  
  # create empty list to return
  cors <- list()
  
  # for each column in one df, compute the correlation and return it
  for(col in seq(1:4)){
    cors[col] <- cor(means_df[,col], factors_df[,col])
  }
  return(cors)
})

cors_df <- rbindlist(cors)
cors_df$value <- values
colnames(cors_df) <- c("far_idf", "far_uniform", 
                  "split_idf", "split_uniform", 
                  "values")
```

Here I plot a heatmap of the correlations:

```{r}
cors_df %>% pivot_longer(
  cols =  c("far_idf", "far_uniform", 
                  "split_idf", "split_uniform"),
  names_to = "model", 
  values_to = "correlation"
  ) %>%
  ggplot(., aes(x=values, y=model, fill = correlation)) +
    geom_tile(color = "white") + 
    geom_text(aes(label = round(correlation, 2)), color = "white")+
  scale_fill_gradient(low = "blue", high = "red") +
    coord_flip() +
  ggtitle("correlation between factor score and mean by setup and value")

rm(cors_df, cors)
```

High correlations for all except Hedonism. 

```{r}
# merge factors and means dataframes 
machines_factors_long <- machines_factors %>% 
  select(item_ID, contains("fa")) %>%
  pivot_longer(cols = contains("fa"), 
               names_sep = "_", 
               names_to = c("value", "model", "weight", "reduction"), values_to = "score")

machines_means_long <- machines_means %>%
  select(item_ID, contains("mean")) %>%
  pivot_longer(cols = contains("mean"), 
               names_sep = "_", 
               names_to = c("value", "model", "weight", "reduction"), values_to = "score")

machines_long <- rbind(machines_means_long, machines_factors_long)

machines_long$setup <- paste(machines_long$model, machines_long$weight, sep = "_")

rm(machines_factors_long, machines_means_long)
```

```{r}
plot_distributions <- function(red, val){
  machines_long %>%
  filter(reduction == red & value == val) %>%
  ggplot(aes(x= score, y = value, fill = model, alpha = 0.7)) +
  facet_grid(~model + weight) +
  geom_density_ridges() + 
  theme(legend.position = "none") + 
  #xlim(-0.12, 0.12) +
  xlab("score") + 
  ylab("") + 
    ggtitle(red)
}
```

```{r}
p1 <- plot_distributions(red = "mean", val = "HEDONISM")
p2 <- plot_distributions(red = "fa", val = "HEDONISM")

p1 / p2
```

```{r}
p3 <- plot_distributions(red = "mean", val = "SECURITY")
p4 <- plot_distributions(red = "fa", val = "SECURITY")

p3 / p4
```


```{r}
machines_factors_long <- machines_factors %>% 
  pivot_longer(cols = -item_ID, 
               names_sep = "_", 
               names_to = c("value", "model", "weight", "reduction"), values_to = "score")

machines_factors_long$setup <- paste(machines_factors_long$model, machines_factors_long$weight)

machines_means_long <- machines_means %>% 
  pivot_longer(cols = -item_ID, 
               names_sep = "_", 
               names_to = c("value", "model", "weight", "reduction"), values_to = "score")

machines_means_long$setup <- paste(machines_means_long$model, machines_means_long$weight)
```


```{r}
p5 <- machines_means_long %>%
  ggplot(aes(x= score, y = setup, fill =value, alpha = 0.7)) +
  facet_grid(~value) +
  geom_density_ridges() + 
  theme(legend.position = "none") + 
  #xlim(-0.12, 0.12) +
  xlab("score") + 
  ylab("model") +
  ggtitle("mean")

p5
```
```{r}
p6 <- machines_factors_long %>%
  ggplot(aes(x= score, y = setup, fill =value, alpha = 0.7)) +
  facet_grid(~value) +
  geom_density_ridges() + 
  theme(legend.position = "none") + 
  #xlim(-0.12, 0.12) +
  xlab("score") + 
  ylab("model") +
  ggtitle("mean")

p6
```
```{r}
rm(p1, p2, p3, p4, p5, p6, cors, cors_df, machines_factors, machines_factors_long, machines_means, machines_means_long, machines_long, plot_distributions)
```


```{r}
build_factor_loadings_df <- function(machines_as_fixed_df, nfactors=1){
  
  # pivot to wide: each machine / value combination is a column
  # x <- values[[3]]
  
  # run factor analyses and store as list
  machine_fas <- lapply(values, function(x){
  machines_as_fixed_df %>% select(starts_with(x)) %>%
    fa(nfactors=nfactors, fm="ml")
  })
  
  # extract factor loadings as a list, format as data.table
  machine_loadings_list <- lapply(machine_fas, function(x){
    df <- x$loadings[1:8,] %>% as.data.frame()
    df$participant_ID <- rownames(df)
    rownames(df) <- NULL
    #colnames(df) <- c("loadings", "participant_ID")
    setDT(df)
    return(df)
  })

  # bind list into a single data.table
  machine_loadings_df <- rbindlist(machine_loadings_list)

  # create a new column for each value
  lapply(values, function(x){
    machine_loadings_df[grepl(x, machine_loadings_df$participant_ID), value := x]
  })

  # remove value name from participant_ID column
  for(value in values){
    machine_loadings_df$participant_ID <- gsub(paste0(value, "_"), "", machine_loadings_df$participant_ID)
  }

  machine_loadings_df$participant_ID <- gsub("_weight_lyrics", "", machine_loadings_df$participant_ID)
  machine_loadings_df$participant_ID <- gsub("mxm_", "", machine_loadings_df$participant_ID)

  return(machine_loadings_df)

}
```

```{r}
plot_factor_loadings <- function(df){
  df %>%
    ggplot(aes(participant_ID, abs(.), fill = .)) +
    facet_wrap(~value, nrow=1) +
    geom_bar(stat = "identity") +
    coord_flip() +
    scale_fill_gradient2(name = "Loading", 
                         high = "blue", mid = "white", low = "red", 
                         midpoint = 0, guide = "none") +
    ylab("loading") +
    theme_bw(base_size =10) +
    theme(axis.text.x= element_text(angle = 90))
}
```


```{r}
factor_1_fa_df <- build_factor_loadings_df(machines_as_fixed_factor_dfs[[4]], nfactors=1)
plot_factor_loadings(factor_1_fa_df)

factor_1_mean_df <- build_factor_loadings_df(machines_as_fixed_mean_dfs[[4]], nfactors=1)
plot_factor_loadings(factor_1_mean_df)
```

```{r}
show_correlation_plot <- function(df, value="POWER"){
  df <- df %>% select(any_of(starts_with(value)))
  cor_mat <- cor(df)
  rownames(cor_mat) <- NULL
  colnames(cor_mat) <- NULL
  corrplot(cor_mat, title = value)
}
```

```{r}
for(value in values){
  show_correlation_plot(machines_as_fixed_factor_dfs[[4]], value=value)
}
```

```{r}
for(value in values){
  show_correlation_plot(machines_as_fixed_mean_dfs[[4]], value=value)
}
```

```{r}
plot_2_factor_loadings <- function(df){
  df <- df %>% pivot_longer(cols = c("ML1", "ML2"), names_to = "factor", values_to = "eigenvalues")
  
  df %>% 
    ggplot(aes(participant_ID, abs(eigenvalues), fill = eigenvalues)) +
    facet_wrap(value~factor, ncol = 4) +
    geom_bar(stat = "identity") +
    coord_flip() +
    scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint = 0, guide = "none") +
    ylab("loading") +
    theme_bw(base_size =10) +
    theme(axis.text.x= element_text(angle = 90))
}
```


```{r}
factor_2_fa_df <- build_factor_loadings_df(machines_as_fixed_factor_dfs[[4]], nfactors=2)

plot_2_factor_loadings(factor_2_fa_df)
```
```{r}
factor_2_mean_df <- build_factor_loadings_df(machines_as_fixed_mean_dfs[[4]], nfactors=2)

plot_2_factor_loadings(factor_2_mean_df)
```


```{r}
factor_2_df <- factor_2_fa_df %>% pivot_longer(cols = c("ML1", "ML2"), names_to = "factor", values_to = "eigenvalues")

factor_2_df %>%
  ggplot(aes(participant_ID, abs(eigenvalues), fill = eigenvalues)) +
  facet_wrap(value~factor, ncol = 4) +
  geom_bar(stat = "identity") +
  coord_flip() +
  scale_fill_gradient2(name = "Loading", 
                       high = "blue", mid = "white", low = "red", 
                       midpoint = 0, guide = "none") +
  ylab("loading") +
  theme_bw(base_size =10) +
  theme(axis.text.x= element_text(angle = 90))
```

Thus it appears we have two distinct estimations: pretrained models vs. mxm data trained models.
